add book details screen components including cover, description, stats, and reviews
