+++ lib/features/book_details/presentation/view/widgets/stat_card.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+class StatCard extends StatelessWidget {
+  final String label;
+  final String value;
+
+  const StatCard({super.key, required this.label, required this.value});
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: Container(
+        padding: const EdgeInsets.all(12),
+        decoration: BoxDecoration(
+          border: Border.all(color: const Color(0xFFDCE0E5)),
+          borderRadius: BorderRadius.circular(8),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              value,
+              style: const TextStyle(
+                color: Color(0xFF111418),
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            const SizedBox(height: 4),
+            Text(
+              label,
+              style: const TextStyle(color: Color(0xFF637488), fontSize: 12),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
